% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\docType{methods}
\name{display,numeric-method}
\alias{display,numeric-method}
\title{Display}
\usage{
\S4method{display}{numeric}(x)
}
\arguments{
\item{x}{MyClass object}
}
\description{
Display method for MyClass objects
}

