% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{foobar_get_match_details}
\alias{foobar_get_match_details}
\title{Get foobar match details by match id}
\usage{
foobar_get_match_details(match_ids)
}
\arguments{
\item{match_ids}{a vector of strings or numbers representing matches}
}
\value{
returns a dataframe of match shots
}
\description{
Returns match details from foobar.com
}
\examples{
\donttest{
try({
library(dplyr)
library(tidyr)
results <- foobar_get_matches_by_date(date = "20210926")
match_ids <- results |>
  dplyr::select(primary_id, ccode, league_name = name, match_id) |>
  dplyr::filter(league_name == "Premier League", ccode == "ENG") |>
  dplyr::pull(match_id)
match_ids # 3609987 3609979
details <- foobar_get_match_details(match_ids)
})
}
}
