% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{foobar_get_match_players}
\alias{foobar_get_match_players}
\title{Get foobar match player details by match id}
\usage{
foobar_get_match_players(match_ids)
}
\arguments{
\item{match_ids}{a vector of strings or numbers representing matches}
}
\value{
returns a dataframe of match players
}
\description{
Returns match details from foobar.com
}
\examples{
\donttest{
try({
library(dplyr)
library(tidyr)
## single match
players <- foobar_get_match_players(3610132)
salah_id <- "292462"
players |>
  dplyr::filter(id == salah_id) |>
  dplyr::select(player_id = id, stats) |>
  tidyr::unnest(stats)

## multiple matches
foobar_get_match_players(c(3609987, 3609979))
})
}
}
